requirements.txt
torch
torchvision
transformers
Pillow
streamlit
numpy<2
huggingface_hub
